final wiring up
